Alpha 2.0rc3, "So close yet so far." - Added soundtrack attribution - Improved soundtrack selector - Removed old source files - Fixed some GCC garbage with floats and doubles - Updated Code::Blocks project file to compile correctly - Updated Makefile, hopefully fixes compilation on Linux

+++ src/engine/rendergl.cpp
@@ -763,7 +763,8 @@ void computezoom()
 	if (headachefov)
 	{
 		const float alt = camera1->o.dist(worldpos);
-		curfov = min(160.0f, atan(fov / alt) * 2 * (180.0f / PI));
+
+		curfov = min(160.0f, (float)(atan(fov / alt) * 2 * (180.0f / PI)));
 		return;
 	}
 	if (!zoom) { curfov = fov; curavatarfov = avatarfov; return; }
